Save predictions from papi2 into personality_entries

+++ database/migrations/2015_07_03_200056_create_personality_table.php
@@ -50,8 +50,8 @@ class CreatePersonalityTable extends Migration {
                        $table->string( 'sys'   );
                        $table->string( 'src'   )->nullable();
                        $table->string( 'name'  )->nullable();
-                       $table->float ( 'value' );
-                       $table->float ( 'error' )->nullable();
+                       $table->double( 'value' );
+                       $table->double( 'error' )->nullable();
                        
                        $table->timestamps();
                        });
